backend/requirements.txt
fastapi==0.95.1
pandas==1.5.3
requests==2.28.2
python-multipart==0.0.6
uvicorn==0.21.1
scikit-learn==1.2.2
pyyaml==6.0
joblib==1.2.0
numpy==1.24.2
implicit==0.6.2
pyarrow==11.0.0
lightgbm==3.3.5
optuna==3.1.1
jsonschema==4.17.3

